Avoid technical jargon in Polish translation
Neither "BTSes" nor "base stations" are legible to regular people. The
only term I found which will be most likely understood correctly is a
"cellular tower". While it's not technically accurate, the English
version uses that and it should be clear enough in this context.

Similarly, "GNSS" is an acronym not commonly known to Poles, while "GPS"
will generally be understood without an issue.

Reviewed-by: Andrzej Masłowski
Reviewed-by: Mariusz Masłowski