WIP: Avoid jargon since the English version avoids it
Instead of calling those "BTSes" or "base stations", I went with just
"cellular transceivers" except for the description of the icon, where
it's a "cellular tower icon" instead, because that's literally what's
in the picture, right? (And don't worry "transceiver" doesn't sound as
technical in Polish as it does in English. Though if you google that
phrase, most of the hits are in the context of microbiology so IDK.)
Another option would be to simply always call them "cellular towers"
(and annoy some people with this inaccuracy).

TODO: Isn't "GNSS" also jargon? But then what about "GPS"?
      (should we consistently use one or the other?)
TODO: "beacon" in Polish could also count as jargon. So if anybody
       can come up with a better word, I'm all ears!
TODO: Squash?